refactor: user dto를 request/response 구조로 분리
